clipboard works (sort-of) in all tracks; multiple text labels are not handled
