+++ src/main/viz/LocationTrack.js
@@ -14,6 +14,7 @@ import canvasUtils from './canvas-utils';
 import dataCanvas from 'data-canvas';
 import style from '../style';
 import d3utils from './d3utils';
+import copyToClipboard from '../Clipboard';
 
 class LocationTrack extends React.Component {
   props: VizProps;
@@ -83,32 +84,8 @@ class LocationTrack extends React.Component {
     ctx.restore();
   }
 
-  copyToClipboard(text) {
-    if (window.clipboardData && window.clipboardData.setData) {
-      // IE specific code path to prevent textarea being shown while dialog is visible.
-      return clipboardData.setData("Text", text);
-    }
-    else if (document.queryCommandSupported && document.queryCommandSupported("copy")) {
-      var textarea = document.createElement("textarea");
-      textarea.textContent = text;
-      textarea.style.position = "fixed";  // Prevent scrolling to bottom of page in MS Edge.
-      document.body.appendChild(textarea);
-      textarea.select();
-      try {
-        return document.execCommand("copy");  // Security exception may be thrown by some browsers.
-      }
-      catch (ex) {
-        console.warn("Copy to clipboard failed.", ex);
-        return false;
-      }
-      finally {
-        document.body.removeChild(textarea);
-      }
-    }
-  }
-
   handleClick(reactEvent: any) {
-    this.copyToClipboard(this.state.midpoint);
+    copyToClipboard(this.state.midpoint);
   }
 }
